refactor(starknet_integration_tests): move fns to sequencer manager mod
commit-id:36d1e190
